+++ CHANGELOG.md
@@ -6,6 +6,21 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 
+## [0.1.3] - 2024-01-30
+### Added
+- PowerShell Profile script.
+- Functions script groups all shared functions into one file to be sourced in profile.
+- Aliases script groups all into one file to be sourced in profile.
+- Prompt script stores the prompt into one space to be sourced whenever the prompt needs overriding.
+- `activate` alias to be used to jump into a python virtual environment quickly.
+- `profile` alias reloads the current profile in the current PowerShell.
+- Meeting - A command to create templated meeting notes.
+- Meeting - KIT (Keeping in Touch) template file.
+- Meeting - KIT notes track information from previous meeting.
+- Meeting - Select from available team members.
+- Meeting - Add new team members option.
+
+
 ## [0.1.2] - 2024-01-18
 ### Added
 - DelEnv - A command to delete a temporary environment created by newenv.
@@ -44,7 +59,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.0] - 2024-01-11
 ### Added
-- DevTools - A command to control and standardise develompment machines.
+- DevTools - A command to control and standardise development machines.
 - DevTools basic setup initialisation.
 - DevTools - Installs [Chocolatey](https://docs.chocolatey.org/en-us/).
 - DevTools - Installs development dependencies.